Move getSequentialName to shared-core
