+++ packages/shared-core/src/helpers/index.ts
@@ -7,3 +7,4 @@ export * as schema from "./schema"
 export * as views from "./views"
 export * as roles from "./roles"
 export * as lists from "./lists"
+export * from "./duplicate"
